+++ src/types/types.ts
@@ -1302,6 +1302,14 @@ export interface GetBaseNftsForContractOptions {
    * Maximum page size is 100.
    */
   pageSize?: number;
+
+  /**
+   * No set timeout by default - When metadata is requested, this parameter is
+   * the timeout (in milliseconds) for the website hosting the metadata to
+   * respond. If you want to only access the cache and not live fetch any
+   * metadata for cache misses then set this value to 0.
+   */
+  tokenUriTimeoutInMs?: number;
 }
 
 /**